Updated 1,2,3,4 to meet teacher's condition
