+++ src/question_1.py
@@ -1,40 +1,19 @@
 #Factorial of a number
-from functools import lru_cache
 
-@lru_cache(maxsize=300)
 def factorial(n):
-    if n == 0:
-        return 1
-    return n* factorial(n-1)
+    a=1
+    for i in range(1,n+1):
+        a*=i
+    return a
+        
 
 #Sum of the first n even numbers
 def sum_first_even(n):
     return (n*(n+1))
 
-
-#alterate
-
-#part 1
-def factorial(a) :
-    b=1
-    while(a!=0) :
-        b*=a
-        a-=1
-    return b
-
-a=int(input("Enter a number : "))
-c=factorial(a)
-print("Required factorial is = ",c)
-
-#part 2
-def sum_of_even(n) :
-    w=0
-    x=0
-    while(w<n) :
-        x+=2*w
-        w+=1
-    return x
+n=int(input("Enter a number: "))
+print(f"Factorial of {n} is: {factorial(n)}")
 
 n=int(input("Enter a number : "))
-y=sum_of_even(n)
+y=sum_first_even(n)
 print("Sum of first ",n," even numbers is = ",y)
