+++ src/question_3/word.txt
@@ -1,0 +1,2 @@
+The text file is a computer file that only contains text and has no special formatting such as bold text, italic text, image, etc.
+With Microsoft Windows computers text files are identified with the .txt file extension.
